+++ lib/home/products/widgets/products_bottom_navigation_bar.dart
@@ -2,7 +2,7 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:m_soko/common/colors.dart';
 import 'package:m_soko/home/products/screens/bottomNavigationItems/profile_page.dart';
-import 'package:m_soko/home/products/widgets/page_transitions.dart';
+import 'package:m_soko/navigation/page_transitions.dart';
 
 class ProductsBottomNavigationBar extends StatefulWidget {
   const ProductsBottomNavigationBar({Key? key}) : super(key: key);
